+++ test/IO/test_io.py
@@ -93,6 +93,7 @@ binary_example_file_list = _convert_filelist([
                                                  'N46E013.hgt',
                                                  'example.zon',
                                                  'example.nc',
+                                                 'example.vk3',
                                                  'example.vk4',
                                                  'example.vk6',
                                                  'mitutoyo_mock.xlsx',
@@ -581,10 +582,12 @@ def test_detect_format(file_format_examples):
     assert detect_format(os.path.join(file_format_examples, 'example-2d.npy')) == 'npy'
     assert detect_format(os.path.join(file_format_examples, 'surface.2048x2048.h5')) == 'h5'
     assert detect_format(os.path.join(file_format_examples, 'example.zon')) == 'zon'
+    assert detect_format(os.path.join(file_format_examples, 'example.vk3')) == 'vk'
     assert detect_format(os.path.join(file_format_examples, 'example.vk4')) == 'vk'
     assert detect_format(os.path.join(file_format_examples, 'example.vk6')) == 'vk'
     assert detect_format(os.path.join(file_format_examples, 'mitutoyo_mock.xlsx')) == 'mitutoyo'
     assert detect_format(os.path.join(file_format_examples, 'mitutoyo_nonuniform_mock.xlsx')) == 'mitutoyo'
+    assert detect_format(os.path.join(file_format_examples, 'example.al3d')) == 'al3d'
     assert detect_format(os.path.join(file_format_examples, 'example_ps.tiff')) == 'ps'