+++ SurfaceTopography/ChangeLog.md
@@ -1,11 +1,12 @@
 Change log for SurfaceTopography
 =============================
 
-v1.2 (not yet released)
------------------------
+v1.2 (14Nov22)
+--------------
 
 - ENH: Reader for Park Systems TIFF
-- ENH: Reader for Keyence VK6
+- ENH: Reader for Keyence VK3 and VK6
+- ENH: Reader for Alicona Imaging AL3D files
 - MAINT: Removed pandas dependency
 - BUG: Fixed bug in VK4 reader